BS-11: moved metric files into metrics subfolders
